(aws) moved stages to aws module
