+++ app/scripts/modules/amazon/aws.module.js
@@ -12,17 +12,17 @@ templates.keys().forEach(function(key) {
 
 module.exports = angular.module('spinnaker.aws', [
   require('../core/cloudProvider/cloudProvider.registry.js'),
-  require('../core/pipeline/config/stages/bake/aws/awsBakeStage.js'),
-  require('../core/pipeline/config/stages/cloneServerGroup/aws/awsCloneServerGroupStage.js'),
-  require('../core/pipeline/config/stages/destroyAsg/aws/awsDestroyAsgStage.js'),
-  require('../core/pipeline/config/stages/disableAsg/aws/awsDisableAsgStage.js'),
-  require('../core/pipeline/config/stages/disableCluster/aws/awsDisableClusterStage.js'),
-  require('../core/pipeline/config/stages/enableAsg/aws/awsEnableAsgStage.js'),
-  require('../core/pipeline/config/stages/findAmi/aws/awsFindAmiStage.js'),
+  require('./pipeline/stages/bake/awsBakeStage.js'),
+  require('./pipeline/stages/cloneServerGroup/awsCloneServerGroupStage.js'),
+  require('./pipeline/stages/destroyAsg/awsDestroyAsgStage.js'),
+  require('./pipeline/stages/disableAsg/awsDisableAsgStage.js'),
+  require('./pipeline/stages/disableCluster/awsDisableClusterStage.js'),
+  require('./pipeline/stages/enableAsg/awsEnableAsgStage.js'),
+  require('./pipeline/stages/findAmi/awsFindAmiStage.js'),
   require('../core/pipeline/config/stages/modifyScalingProcess/modifyScalingProcess.module.js'),
-  require('../core/pipeline/config/stages/resizeAsg/aws/awsResizeAsgStage.js'),
-  require('../core/pipeline/config/stages/scaleDownCluster/aws/awsScaleDownClusterStage.js'),
-  require('../core/pipeline/config/stages/shrinkCluster/aws/awsShrinkClusterStage.js'),
+  require('./pipeline/stages/resizeAsg/awsResizeAsgStage.js'),
+  require('./pipeline/stages/scaleDownCluster/awsScaleDownClusterStage.js'),
+  require('./pipeline/stages/shrinkCluster/awsShrinkClusterStage.js'),
   require('../core/subnet/subnet.module.js'),
   require('./serverGroup/details/serverGroup.details.module.js'),
   require('./serverGroup/serverGroup.transformer.js'),
