+++ app/scripts/modules/amazon/pipeline/stages/bake/bakeExecutionDetails.html
