+++ app/scripts/modules/amazon/pipeline/stages/bake/bakeStage.html
