+++ app/scripts/modules/amazon/pipeline/stages/cloneServerGroup/awsCloneServerGroupStage.js
@@ -3,9 +3,9 @@
 let angular = require('angular');
 
 module.exports = angular.module('spinnaker.core.pipeline.stage.aws.cloneServerGroupStage', [
-  require('../../../../../application/modal/platformHealthOverride.directive.js'),
-  require('../../../../../account/account.service.js'),
-  require('../../../../../../core/naming/naming.service.js'),
+  require('../../../../core/application/modal/platformHealthOverride.directive.js'),
+  require('../../../../core/account/account.service.js'),
+  require('../../../../core/naming/naming.service.js'),
   require('./cloneServerGroupExecutionDetails.controller.js'),
 ])
   .config(function(pipelineConfigProvider) {
