+++ app/scripts/modules/amazon/pipeline/stages/cloneServerGroup/cloneServerGroupExecutionDetails.html
