+++ app/scripts/modules/amazon/pipeline/stages/cloneServerGroup/cloneServerGroupStage.html
