+++ app/scripts/modules/amazon/pipeline/stages/cloneServerGroup/cloneServerGroupStepLabel.html
