+++ app/scripts/modules/amazon/pipeline/stages/destroyAsg/destroyAsgExecutionDetails.html
