+++ app/scripts/modules/amazon/pipeline/stages/destroyAsg/destroyAsgStage.html
