+++ app/scripts/modules/amazon/pipeline/stages/destroyAsg/destroyAsgStepLabel.html
