+++ app/scripts/modules/amazon/pipeline/stages/disableAsg/awsDisableAsgStage.js
@@ -1,31 +1,32 @@
 'use strict';
 
-//BEN_TODO: where is this defined?
-
 let angular = require('angular');
 
-module.exports = angular.module('spinnaker.core.pipeline.stage.aws.enableAsgStage', [
-  require('../../../../../application/modal/platformHealthOverride.directive.js'),
-  require('../../stageConstants.js'),
-  require('./enableAsgExecutionDetails.controller.js')
+module.exports = angular.module('spinnaker.core.pipeline.stage.aws.disableAsgStage', [
+  require('../../../../core/application/modal/platformHealthOverride.directive.js'),
+  require('../../../../core/pipeline/config/stages/stageConstants.js'),
+  require('./disableAsgExecutionDetails.controller.js')
 ])
   .config(function(pipelineConfigProvider) {
     pipelineConfigProvider.registerStage({
-      provides: 'enableServerGroup',
-      alias: 'enableAsg',
+      provides: 'disableServerGroup',
+      alias: 'disableAsg',
       cloudProvider: 'aws',
-      templateUrl: require('./enableAsgStage.html'),
-      executionDetailsUrl: require('./enableAsgExecutionDetails.html'),
-      executionStepLabelUrl: require('./enableAsgStepLabel.html'),
+      templateUrl: require('./disableAsgStage.html'),
+      executionDetailsUrl: require('./disableAsgExecutionDetails.html'),
+      executionStepLabelUrl: require('./disableAsgStepLabel.html'),
       validators: [
+        {
+          type: 'targetImpedance',
+          message: 'This pipeline will attempt to disable a server group without deploying a new version into the same cluster.'
+        },
         { type: 'requiredField', fieldName: 'cluster' },
         { type: 'requiredField', fieldName: 'target', },
         { type: 'requiredField', fieldName: 'regions', },
         { type: 'requiredField', fieldName: 'credentials', fieldLabel: 'account'},
       ],
     });
-  }).controller('awsEnableAsgStageCtrl', function($scope, accountService, stageConstants) {
-    var ctrl = this;
+  }).controller('awsDisableAsgStageCtrl', function($scope, accountService, stageConstants) {
 
     let stage = $scope.stage;
 
@@ -39,10 +40,6 @@ module.exports = angular.module('spinnaker.core.pipeline.stage.aws.enableAsgStag
       $scope.state.accounts = true;
     });
 
-    ctrl.reset = () => {
-      ctrl.accountUpdated();
-      ctrl.resetSelectedCluster();
-    };
 
     $scope.targets = stageConstants.targetList;
 
@@ -64,6 +61,5 @@ module.exports = angular.module('spinnaker.core.pipeline.stage.aws.enableAsgStag
       stage.target = $scope.targets[0].val;
     }
 
-    $scope.$watch('stage.credentials', $scope.accountUpdated);
   });
