+++ app/scripts/modules/amazon/pipeline/stages/disableAsg/disableAsgExecutionDetails.html
