+++ app/scripts/modules/amazon/pipeline/stages/disableAsg/disableAsgStage.html
