+++ app/scripts/modules/amazon/pipeline/stages/disableAsg/disableAsgStepLabel.html
