+++ app/scripts/modules/amazon/pipeline/stages/disableCluster/awsDisableClusterStage.js
