+++ app/scripts/modules/amazon/pipeline/stages/disableCluster/disableClusterExecutionDetails.html
