+++ app/scripts/modules/amazon/pipeline/stages/disableCluster/disableClusterStage.html
