+++ app/scripts/modules/amazon/pipeline/stages/enableAsg/awsEnableAsgStage.js
@@ -1,32 +1,31 @@
 'use strict';
 
+//BEN_TODO: where is this defined?
+
 let angular = require('angular');
 
-module.exports = angular.module('spinnaker.core.pipeline.stage.aws.disableAsgStage', [
-  require('../../../../../application/modal/platformHealthOverride.directive.js'),
-  require('../../stageConstants.js'),
-  require('./disableAsgExecutionDetails.controller.js')
+module.exports = angular.module('spinnaker.core.pipeline.stage.aws.enableAsgStage', [
+  require('../../../../core/application/modal/platformHealthOverride.directive.js'),
+  require('../../../../core/pipeline/config/stages/stageConstants.js'),
+  require('./enableAsgExecutionDetails.controller.js')
 ])
   .config(function(pipelineConfigProvider) {
     pipelineConfigProvider.registerStage({
-      provides: 'disableServerGroup',
-      alias: 'disableAsg',
+      provides: 'enableServerGroup',
+      alias: 'enableAsg',
       cloudProvider: 'aws',
-      templateUrl: require('./disableAsgStage.html'),
-      executionDetailsUrl: require('./disableAsgExecutionDetails.html'),
-      executionStepLabelUrl: require('./disableAsgStepLabel.html'),
+      templateUrl: require('./enableAsgStage.html'),
+      executionDetailsUrl: require('./enableAsgExecutionDetails.html'),
+      executionStepLabelUrl: require('./enableAsgStepLabel.html'),
       validators: [
-        {
-          type: 'targetImpedance',
-          message: 'This pipeline will attempt to disable a server group without deploying a new version into the same cluster.'
-        },
         { type: 'requiredField', fieldName: 'cluster' },
         { type: 'requiredField', fieldName: 'target', },
         { type: 'requiredField', fieldName: 'regions', },
         { type: 'requiredField', fieldName: 'credentials', fieldLabel: 'account'},
       ],
     });
-  }).controller('awsDisableAsgStageCtrl', function($scope, accountService, stageConstants) {
+  }).controller('awsEnableAsgStageCtrl', function($scope, accountService, stageConstants) {
+    var ctrl = this;
 
     let stage = $scope.stage;
 
@@ -40,6 +39,10 @@ module.exports = angular.module('spinnaker.core.pipeline.stage.aws.disableAsgSta
       $scope.state.accounts = true;
     });
 
+    ctrl.reset = () => {
+      ctrl.accountUpdated();
+      ctrl.resetSelectedCluster();
+    };
 
     $scope.targets = stageConstants.targetList;
 
@@ -61,5 +64,6 @@ module.exports = angular.module('spinnaker.core.pipeline.stage.aws.disableAsgSta
       stage.target = $scope.targets[0].val;
     }
 
+    $scope.$watch('stage.credentials', $scope.accountUpdated);
   });
