+++ app/scripts/modules/amazon/pipeline/stages/enableAsg/enableAsgStage.html
