+++ app/scripts/modules/amazon/pipeline/stages/enableAsg/enableAsgStepLabel.html
