+++ app/scripts/modules/amazon/pipeline/stages/findAmi/awsFindAmiStage.js
