+++ app/scripts/modules/amazon/pipeline/stages/findAmi/findAmiExecutionDetails.html
