+++ app/scripts/modules/amazon/pipeline/stages/findAmi/findAmiStage.html
