+++ app/scripts/modules/amazon/pipeline/stages/resizeAsg/awsResizeAsgStage.js
@@ -3,7 +3,7 @@
 let angular = require('angular');
 
 module.exports = angular.module('spinnaker.core.pipeline.stage.aws.resizeAsgStage', [
-  require('../../../../../application/modal/platformHealthOverride.directive.js'),
+  require('../../../../core/application/modal/platformHealthOverride.directive.js'),
   require('./resizeAsgExecutionDetails.controller.js'),
 ])
   .config(function(pipelineConfigProvider) {
