+++ app/scripts/modules/amazon/pipeline/stages/resizeAsg/resizeAsgExecutionDetails.html
