+++ app/scripts/modules/amazon/pipeline/stages/resizeAsg/resizeAsgStage.html
