+++ app/scripts/modules/amazon/pipeline/stages/resizeAsg/resizeAsgStepLabel.html
