+++ app/scripts/modules/amazon/pipeline/stages/scaleDownCluster/awsScaleDownClusterStage.js
