+++ app/scripts/modules/amazon/pipeline/stages/scaleDownCluster/scaleDownClusterExecutionDetails.controller.js
@@ -4,8 +4,8 @@ let angular = require('angular');
 
 module.exports = angular.module('spinnaker.core.pipeline.stage.scaleDownCluster.aws.executionDetails.controller', [
     require('angular-ui-router'),
-    require('../../../../../delivery/details/executionDetailsSection.service.js'),
-    require('../../../../../delivery/details/executionDetailsSectionNav.directive.js'),
+    require('../../../../core/delivery/details/executionDetailsSection.service.js'),
+    require('../../../../core/delivery/details/executionDetailsSectionNav.directive.js'),
 ])
   .controller('awsScaleDownClusterExecutionDetailsCtrl', function ($scope, $stateParams, executionDetailsSectionService) {
