+++ app/scripts/modules/amazon/pipeline/stages/scaleDownCluster/scaleDownClusterExecutionDetails.html
