+++ app/scripts/modules/amazon/pipeline/stages/scaleDownCluster/scaleDownClusterStage.html
