+++ app/scripts/modules/amazon/pipeline/stages/shrinkCluster/shrinkClusterExecutionDetails.html
