+++ app/scripts/modules/amazon/pipeline/stages/shrinkCluster/shrinkClusterStage.html
